refactor: Header 컴포넌트 개선 - 재사용 가능한 컴포넌트 분리(로고, 아바타 등)
