+++ src/components/Avatar/styles.module.scss
@@ -1,0 +1,10 @@
+.avatar {
+  width: 4rem;
+  height: 4rem;
+  border-radius: 50%;
+  overflow: hidden;
+
+  &:hover {
+    outline: 2px solid var(--color-secondary-100);
+  }
+}
